Random values generator moved to mock mod
