+++ memcrs/src/mock/mod.rs
@@ -1,2 +1,3 @@
 pub mod mock_server;
 pub mod value;
+pub mod key_value;
